+++ packages/quantic/force-app/main/default/lwc/quanticNumericFacet/quanticNumericFacet.js
@@ -3,7 +3,8 @@ import {
   registerComponentForInit,
   initializeWithHeadless,
   getHeadlessBindings,
-  registerToStore
+  registerToStore,
+  getHeadlessBundle,
 } from 'c/quanticHeadlessLoader';
 import {I18nUtils, Store} from 'c/quanticUtils';
 import LOCALE from '@salesforce/i18n/locale';
@@ -12,12 +13,12 @@ import clearFilter from '@salesforce/label/c.quantic_ClearFilter';
 import clearFilter_plural from '@salesforce/label/c.quantic_ClearFilter_plural';
 import collapseFacet from '@salesforce/label/c.quantic_CollapseFacet';
 import expandFacet from '@salesforce/label/c.quantic_ExpandFacet';
-import min from "@salesforce/label/c.quantic_Min";
-import max from "@salesforce/label/c.quantic_Max";
-import numberInputMinimum from "@salesforce/label/c.quantic_NumberInputMinimum";
-import numberInputMaximum from "@salesforce/label/c.quantic_NumberInputMaximum";
-import apply from "@salesforce/label/c.quantic_Apply";
-import numberInputApply from "@salesforce/label/c.quantic_NumberInputApply";
+import min from '@salesforce/label/c.quantic_Min';
+import max from '@salesforce/label/c.quantic_Max';
+import numberInputMinimum from '@salesforce/label/c.quantic_NumberInputMinimum';
+import numberInputMaximum from '@salesforce/label/c.quantic_NumberInputMaximum';
+import apply from '@salesforce/label/c.quantic_Apply';
+import numberInputApply from '@salesforce/label/c.quantic_NumberInputApply';
 import messageWhenRangeOverflow from '@salesforce/label/c.quantic_MessageWhenRangeOverflow';
 import messageWhenRangeUnderflow from '@salesforce/label/c.quantic_MessageWhenRangeUnderflow';
 
@@ -32,6 +33,7 @@ import messageWhenRangeUnderflow from '@salesforce/label/c.quantic_MessageWhenRa
 /**
  * The `QuanticNumericFacet` component displays facet values as numeric ranges.
  * @category Search
+ * @category Insight Panel
  * @example
  * <c-quantic-numeric-facet engine-id={engineId} facet-id="myfacet" field="ytlikecount" label="Youtube Likes" numberOfValues="5" sort-criteria="descending" range-algorithm="even" formatting-function={myFormattingFunction} is-collapsed></c-quantic-numeric-facet>
  */
@@ -42,7 +44,7 @@ export default class QuanticNumericFacet extends LightningElement {
    * @type {string}
    */
   @api engineId;
-  /** 
+  /**
    * A unique identifier for the facet.
    * @api
    * @type {string}
@@ -94,11 +96,10 @@ export default class QuanticNumericFacet extends LightningElement {
    * @param {NumericFacetValue} item
    * @returns {string}
    */
-  @api formattingFunction = (item) => `${new Intl.NumberFormat(LOCALE).format(
-    item.start
-  )} - ${new Intl.NumberFormat(LOCALE).format(
-    item.end
-  )}`;
+  @api formattingFunction = (item) =>
+    `${new Intl.NumberFormat(LOCALE).format(
+      item.start
+    )} - ${new Intl.NumberFormat(LOCALE).format(item.end)}`;
   /**
    * Whether this facet should contain an input allowing users to set custom ranges.
    * Depending on the field, the input can allow either decimal or integer values.
@@ -131,7 +132,7 @@ export default class QuanticNumericFacet extends LightningElement {
     'sortCriteria',
     'rangeAlgorithm',
     'withInput',
-  ]
+  ];
 
   /** @type {NumericFacetState} */
   @track state;
@@ -139,8 +140,8 @@ export default class QuanticNumericFacet extends LightningElement {
   @track filterState = {
     isLoading: false,
     enabled: true,
-    facetId: undefined
-  }
+    facetId: undefined,
+  };
 
   /** @type {NumericFacet} */
   facet;
@@ -156,6 +157,8 @@ export default class QuanticNumericFacet extends LightningElement {
   unsubscribeFilter;
   /** @type {Function} */
   unsubscribeSearchStatus;
+  /** @type {AnyHeadless} */
+  headless;
 
   /** @type {string} */
   start;
@@ -178,7 +181,7 @@ export default class QuanticNumericFacet extends LightningElement {
     apply,
     numberInputApply,
     messageWhenRangeOverflow,
-    messageWhenRangeUnderflow
+    messageWhenRangeUnderflow,
   };
 
   connectedCallback() {
@@ -193,32 +196,35 @@ export default class QuanticNumericFacet extends LightningElement {
    * @param {SearchEngine} engine
    */
   initialize = (engine) => {
-    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
+    this.headless = getHeadlessBundle(this.engineId);
+    this.searchStatus = this.headless.buildSearchStatus(engine);
     this.unsubscribeSearchStatus = this.searchStatus.subscribe(() =>
       this.updateState()
     );
 
-    if(this.numberOfValues > 0) {
+    if (this.numberOfValues > 0) {
       this.initializeFacet(engine);
     }
-    if(this.withInput) {
+    if (this.withInput) {
       this.initializeFilter(engine);
     }
-    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
-    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() => this.updateState());
+    this.searchStatus = this.headless.buildSearchStatus(engine);
+    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() =>
+      this.updateState()
+    );
     registerToStore(this.engineId, Store.facetTypes.NUMERICFACETS, {
       label: this.label,
       facetId: this.facet?.state.facetId ?? this.field,
       format: this.formattingFunction,
-      element: this.template.host
+      element: this.template.host,
     });
-  }
+  };
 
   /**
    * @param {import("coveo").SearchEngine} engine
    */
   initializeFacet(engine) {
-    this.facet = CoveoHeadless.buildNumericFacet(engine, {
+    this.facet = this.headless.buildNumericFacet(engine, {
       options: {
         field: this.field,
         generateAutomaticRanges: true,
@@ -226,23 +232,26 @@ export default class QuanticNumericFacet extends LightningElement {
         rangeAlgorithm: this.rangeAlgorithm,
         numberOfValues: Number(this.numberOfValues),
         facetId: this.facetId ?? this.field,
-      }
+      },
     });
     this.unsubscribe = this.facet.subscribe(() => this.updateState());
-    
   }
 
   /**
    * @param {import("coveo").SearchEngine} engine
    */
   initializeFilter(engine) {
-    this.numericFilter = CoveoHeadless.buildNumericFilter(engine, {
+    this.numericFilter = this.headless.buildNumericFilter(engine, {
       options: {
         field: this.field,
-        facetId: this.facet?.state.facetId ? `${this.facet.state.facetId}_input` : undefined
-      }
+        facetId: this.facet?.state.facetId
+          ? `${this.facet.state.facetId}_input`
+          : undefined,
+      },
     });
-    this.unsubscribeFilter = this.numericFilter.subscribe(() => this.updateFilterState());
+    this.unsubscribeFilter = this.numericFilter.subscribe(() =>
+      this.updateFilterState()
+    );
   }
 
   disconnectedCallback() {
@@ -253,7 +262,10 @@ export default class QuanticNumericFacet extends LightningElement {
 
   updateState() {
     this.state = this.facet?.state;
-    this.showPlaceholder = this.searchStatus?.state?.isLoading && !this.searchStatus?.state?.hasError && !this.searchStatus?.state?.firstSearchExecuted;
+    this.showPlaceholder =
+      this.searchStatus?.state?.isLoading &&
+      !this.searchStatus?.state?.hasError &&
+      !this.searchStatus?.state?.firstSearchExecuted;
   }
 
   updateFilterState() {
@@ -277,7 +289,7 @@ export default class QuanticNumericFacet extends LightningElement {
   }
 
   get step() {
-    return this.withInput ==='integer' ? '1' : 'any';
+    return this.withInput === 'integer' ? '1' : 'any';
   }
 
   /** @returns {HTMLInputElement} */
@@ -303,28 +315,45 @@ export default class QuanticNumericFacet extends LightningElement {
   }
 
   get actionButtonLabel() {
-    const label = this.isCollapsed ? this.labels.expandFacet : this.labels.collapseFacet;
+    const label = this.isCollapsed
+      ? this.labels.expandFacet
+      : this.labels.collapseFacet;
     return I18nUtils.format(label, this.label);
   }
 
   get numberOfSelectedValues() {
-    return this.filterState?.range ? 1 : this.state.values.filter(({state}) => state === 'selected').length;
+    return this.filterState?.range
+      ? 1
+      : this.state.values.filter(({state}) => state === 'selected').length;
   }
 
   get showValues() {
-    return !this.searchStatus?.state?.hasError && !this.filterState?.range && !!this.values.length;
+    return (
+      !this.searchStatus?.state?.hasError &&
+      !this.filterState?.range &&
+      !!this.values.length
+    );
   }
 
   get clearFilterLabel() {
     if (this.hasActiveValues) {
-      const labelName = I18nUtils.getLabelNameWithCount('clearFilter', this.numberOfSelectedValues);
-      return `${I18nUtils.format(this.labels[labelName], this.numberOfSelectedValues)}`;
+      const labelName = I18nUtils.getLabelNameWithCount(
+        'clearFilter',
+        this.numberOfSelectedValues
+      );
+      return `${I18nUtils.format(
+        this.labels[labelName],
+        this.numberOfSelectedValues
+      )}`;
     }
     return '';
   }
 
   get shouldRenderInput() {
-    return this.withInput && this.searchStatus?.state?.hasResults || !!this.filterState?.range;
+    return (
+      (this.withInput && this.searchStatus?.state?.hasResults) ||
+      !!this.filterState?.range
+    );
   }
 
   get shouldRenderValues() {
@@ -350,22 +379,24 @@ export default class QuanticNumericFacet extends LightningElement {
     this.inputMax.required = false;
   }
 
-  /** 
+  /**
    * @param {CustomEvent<{value: string}>} evt
    */
   onSelectValue(evt) {
-    const item = this.values.find((value) => this.formattingFunction(value) === evt.detail.value);
+    const item = this.values.find(
+      (value) => this.formattingFunction(value) === evt.detail.value
+    );
     this.facet.toggleSelect(item);
   }
 
   clearSelections() {
-    if(this.filterState?.range) {
+    if (this.filterState?.range) {
       this.numericFilter.clear();
     }
     this.facet?.deselectAll();
-    if(this.withInput) {
+    if (this.withInput) {
       this.resetValidityParameters();
-    
+
       this.allInputs.forEach((input) => {
         // @ts-ignore
         input.checkValidity();
@@ -385,27 +416,34 @@ export default class QuanticNumericFacet extends LightningElement {
 
   /**
    * @param {Event} evt
-  */
+   */
   onApply(evt) {
     evt.preventDefault();
 
     this.setValidityParameters();
-    // @ts-ignore
-    const allValid = this.allInputs.reduce((validSoFar, inputCmp) => validSoFar && inputCmp.reportValidity(), true);
+    const allValid = this.allInputs.reduce(
+      // @ts-ignore
+      (validSoFar, inputCmp) => validSoFar && inputCmp.reportValidity(),
+      true
+    );
     this.resetValidityParameters();
 
     if (!allValid) {
       return;
     }
 
-    if(this.numberOfValues > 0) {
+    if (this.numberOfValues > 0) {
       const engine = getHeadlessBindings(this.engineId).engine;
-      engine.dispatch(CoveoHeadless.loadNumericFacetSetActions(engine).deselectAllNumericFacetValues(this.facet.state.facetId));
+      engine.dispatch(
+        this.headless
+          .loadNumericFacetSetActions(engine)
+          .deselectAllNumericFacetValues(this.facet.state.facetId)
+      );
     }
-    
+
     this.numericFilter.setRange({
       start: this.inputMin ? Number(this.inputMin.value) : undefined,
-      end: this.inputMax ? Number(this.inputMax.value) : undefined
+      end: this.inputMax ? Number(this.inputMax.value) : undefined,
     });
   }
 
@@ -445,7 +483,7 @@ export default class QuanticNumericFacet extends LightningElement {
   get customMessageOverflow() {
     return I18nUtils.format(this.labels.messageWhenRangeOverflow, this.max);
   }
-  
+
   get customMessageUnderflow() {
     return I18nUtils.format(this.labels.messageWhenRangeUnderflow, this.min);
   }