+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.90.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.90.0...@coveo/atomic@1.90.1) (2022-07-27)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.90.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.89.2...@coveo/atomic@1.90.0) (2022-07-27)
+
+
+### Features
+
+* **headless:** Refactor core and insight controllers ([#2215](https://github.com/coveo/ui-kit/issues/2215)) ([25fb657](https://github.com/coveo/ui-kit/commit/25fb657e43c725beb83a48a9e5fc8393a649c051))
+
+
+
+
+
+## [1.89.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.89.1...@coveo/atomic@1.89.2) (2022-07-26)
+
+
+### Bug Fixes
+
+* **atomic:** added accessible group for results per page options ([#2245](https://github.com/coveo/ui-kit/issues/2245)) ([a08b855](https://github.com/coveo/ui-kit/commit/a08b85545fcdb30a98575bb1e381d6ad057a0f4b))
+* **atomic:** focus first new value in facets when pressing show more ([#2250](https://github.com/coveo/ui-kit/issues/2250)) ([90f68d1](https://github.com/coveo/ui-kit/commit/90f68d1634014630545168b7a03f10b826f7dd90))
+
+
+
+
+
+## [1.89.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.89.0...@coveo/atomic@1.89.1) (2022-07-26)
+
+
+### Bug Fixes
+
+* **atomic:** use custom "closest" method for initializeBindings, resultContext ([#2242](https://github.com/coveo/ui-kit/issues/2242)) ([d5194a0](https://github.com/coveo/ui-kit/commit/d5194a0e622cc15e7a63ef74c5fe2656054bd76e))
+
+
+
+
+
+# [1.89.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.88.2...@coveo/atomic@1.89.0) (2022-07-25)
+
+
+### Bug Fixes
+
+* **atomic:** improved aria label for breadcrumbs ([#2243](https://github.com/coveo/ui-kit/issues/2243)) ([3073425](https://github.com/coveo/ui-kit/commit/3073425a06d0e74bf4a0a9ce413141018cd097a5))
+
+
+### Features
+
+* **atomic:** added "-selected" parts for link and box facet values ([#2236](https://github.com/coveo/ui-kit/issues/2236)) ([79f51b4](https://github.com/coveo/ui-kit/commit/79f51b49a6897d6f9cc22da196747c9b83f70cec))
+
+
+
+
+
+## [1.88.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.88.1...@coveo/atomic@1.88.2) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
 ## [1.88.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.88.0...@coveo/atomic@1.88.1) (2022-07-25)
 
 **Note:** Version bump only for package @coveo/atomic